split events feed in own file
